requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
astunparse==1.6.3
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
prometheus_client==0.21.0
protobuf==4.25.5
psycopg2==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-jose==3.3.0
python-multipart==0.0.10
requests==2.32.3
rich==13.8.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wrapt==1.16.0
